Create CGPA Calculator
In this program, the student has to enter some basic details such as no. of courses taken, credits of each course and grades earned in each course. CGPA Calculator calculates total credits, total grades of each semester and determines GPA for the semester. Based on all this data, it displays the CGPA of the student.